requirements.txt
fastapi
uvicorn[standard]
aiogram
sqlalchemy
alembic
psycopg[binary]
psycopg2-binary
celery[redis]
redis
httpx
pydantic
pydantic-settings
structlog
boto3
python-dateutil
python-multipart
Pillow
google-genai>=1.58.0
numpy
pytest
pytest-asyncio
respx
cryptography
argon2-cffi
jinja2
